backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.1.0
